UNR-543 Fix default map loading (#307)
#### Description
Fixing a bug which caused the PIE clients booted from the editor to always use the GameDefaultMap and ignore the currently open man. This did not match Unreal's default behaviour and made map switching a pain. This could also lead to strange behaviour where the server loads one map and the clients another.
#### Tests
Tested by changing the GameDefaultMap / ServerDefaultMap and ensuring that the currently open map is used for both the server and the clients.
What automated tests are included in this PR?
#### Documentation
There is a known issue which can be removed.
#### Primary reviewers
@m-samiec @girayimprobable