Bug fix - removing from a set whilst iterating over it.
For `ResetOutgoingUpdate`. make a list of invalid objects, to remove after iterating on the set.

For `CreateRPCCommandRequest` and `CreateMulticastRequest`, `UnresolvedObjects.Remove(Object);` is useless here because the set is only used to get the first element then return.

For `QueueOutgoingUpdate`, it is not expected to ever have invalid subobjects, so the remove has been removed, but if it is ever found to be reached, we should deal with it properly.