[UNR-861] URL locator param renamed and cleanup
Renamed URL option locator to legacylocator, and renamed locatorv2 to
locator.

Properly destroy the locator in DestoryConnection.

Changed error logging to use the same function as the other
Receptionist/LegacyLocator code paths.

Checked for the playerIdentityToken instead of loginToken to properly
tell the difference between the legacy and non legacy locators.