UNR-937 Load Default GameMode for schema generation (#668)
**Contributions**: We are not currently taking public contributions - see our [contributions](CONTRIBUTING.md) policy. However, we are accepting issues and we do want your [feedback](../README.md#give-us-feedback).

-------

#### Description
In the situation that you have selected a default GameMode and not set an override, your default GameMode will not be loaded at bootup until you access the asset or load your project settings. This meant that schema would not be generated for this GameMode resulting in a runtime crash. This fix simply force loads the GlobalDefaultGameMode before generating schema ensuring it is in memory for generation.

#### Release note
Bugfix: Fixed an issue with schema generation for the default GameMode.

#### Tests
Tested on TPS with DefaultGameMode set and no map override, without fix we have a runtime crash. With fix the GameMode gets scheme generated and no crash.

#### Documentation
Known issue needs removing

#### Primary reviewers
@Vatyx @m-samiec