[UNR-1641] - Spatial Position not updating for AI Characters (#1111)
* [UNR-1641] - Spatial Position not updating for AI Characters

#### Description
AI Characters were not having their Spatial position updated in the inspector, making it difficult to know which entity in the inspector corresponded to which Character in game. Actor replication in SpatialActorChannel would not update an Actor's Spatial position if it had a parent and the Actor itself was not an `APlayerController`.

Fixed by adding an extra check to `UpdateSpatialPosition`, if the replicating Actor has an owner and that owner is replicated, return early as before since the owner will be replicated and have it's Spatial position updated along with all of its children. In the case that the replicating Actor's owner is *not* replicated, continue and update its Spatial position

#### Release note
Fixed issue where an Actor's Spatial position was not updated if it had an owner that was not replicated (eg. AI Character possessed by an AIController)

#### Primary reviewers
@mattyoung-improbable @improbable-valentyn

* Update CHANGELOG.md

* Update SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialActorChannel.cpp

Co-Authored-By: improbable-valentyn <32096431+improbable-valentyn@users.noreply.github.com>

* Update SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialActorChannel.cpp

Co-Authored-By: improbable-valentyn <32096431+improbable-valentyn@users.noreply.github.com>

* Update SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialActorChannel.cpp

Co-Authored-By: improbable-valentyn <32096431+improbable-valentyn@users.noreply.github.com>