Home page and all Get Started fix-ups
https://improbableio.atlassian.net/browse/DOC-1276
Sub edit home page.

Sub edit Get started intro page. (Update with new functionality.)

Auto-install and Manual-install: - Add links to and from the manual install for 2. Get the fork, and 3.1 Example project, 3.1 Starter Template. - Make sure the user orientation makes sense between them and add the manual-install pages to the Table of Contents with relevant user orientation introduction.

Update 2. Get the fork to be 2. Get the fork and the plugin. Standardise page name and page content.

Make sure the new and updated content in the Get Started Example Project and Get Started Starter Template:  - follows a linear narrative and exposition (we explain concepts and products when we first reference them - the Console, SpatialOS Runtime, ++). - Is not confusing to follow  - use formatting to space out the steps. - the proliferation of warnings are incorporated into the text so they aren’t jarring. - the lost explanations  are back in (you need to invent an assembly name and deployment name) - the instructions are written as instructions (do x, to do this, do y) - the style is consistent (select x, the GDK toolbar, native Unreal networking, it’s all imperative and active voice with no gerunds (especially in the titles).) - Screen grabs and instructions are consistent with the project and current UI. - There are no term regressions (for exampele go from “server-worker” back to “server”) or synonyms (we only use “game client” and don’t flip between “game client” and “client-worker”).

Standardise fork and plugin terms.

Fix the BuildWorkers.bat instructions so it’s clear that you use different paths from the command line, depending on whether you used  - do this for Example Project, Starter Template, Multiserver Zoning.