Fix linux build, and cloud deployments
Worker_RequestId now has the same problem that Worker_EntityId had,
where the type it resolves to is not hashable in a TMap. Fixed some
additional library link dependencies, and corrected the connector flow,
which now requires the SpatialOS project name - we can read this from
the command line arguments.