Add ActorClass constraint to ComponentConstraint
Fixes UNR-2426 but also requires an ActorClass constraint on a
ComponentConstraint.

Currently the ComponentConstraint looked up the Component ClassPath
using the map for Actor ClassPaths, this would never return a
Component ID for a component. Now this uses the ActorClass to lookup the
generated component for the ActorComponent.

The ActorClass is needed as the GDK generates new components for
each Actor when an ActorComponent is attached.