Fixing two issues.
There was a race where the server had loaded the map, the GSM had been created but the server did not have authority over the SpatialConstants::STARTUP_ACTOR_MANAGER_COMPONENT_ID component on the GSM.
This lead to a failed check() in SetCanBeginPlay.

Also adding the !GlobalStateManager->IsReadyToCallBeginPlay() check.
This is just in case we load the map after we gain authority over the component on the GSM. When we gain autority the GSM has will have already called SetCanBeginPlay(true) so we shouldn't do this again.

Also removing th log as I don't think it is helpful. The flow of not doing the GlobalStateManager->SetCanBeginPlay(true) logic is perfectly valid. Essentially we only want to go down this path on server travel.