Adds IndexFromSchema functions for various classes
As part of the Snapshot Migration work, we need to extract existing
SchemaObject values from SchemaObjects representing the data on a
component. While many of the Unreal-facing classes that represent
SchemaObjects do have IndexFromSchema functions (which facilitates
extracting collections from SchemaObjects) defined, the ones addressed
in this PR did not.

The Migrator doesn't reason about cardinality of a
field in Schema, since the Singular and Optional field cases are simply
degenerate Arrays (with 1 and 0 or 1 elements, respectively). As such,
it assumes everything can be treated as an array and maintains a thin
compatability wrapper that allows us to use IndexFromSchema functions to
extract an arbitrary number of _correctly-typed_ objects from a
SchemaObject; this is important as we may need to apply a transform (e.g., updating component Ids to accomodate schema changes) to the object before writing it back to the new snapshot.