Extracts & encapsulates Actor->Entity logic
As part of Snapshot Migration we go through most of the same process that the
SpatialSender does when initially creating an Entity; we take an Actor
(well, actually its Channel) and enumerate the Spatial Components that
represent that Actor as an Entity.

The Sender follows this up by sending a CreateEntity request while the
Migrator takes the resulting Components and "fast-forwards" data from
the old Components (from the existing Entity that we read out of the
Snapshot) onto them when possible and where appropriate.

This change just pulls the logic for actually creating a list of
Components from an Actor out into its own encapsulated thing that can be
invoked without needing to route through the SpatialSender; right now
everything is still shoved into one function but I think breaking it
down into a number of smaller, more easily parseable functions that are
invoked internally would be a low-hanging next step.