Feature/UNR-2589 latency tracking (#1717)
These changes add a static function to the existing latency tracing code that allows to easily set the prefixes of trace names. This facilitates the use of the new latency tracing scenario, aa users are expected to specify a trace prefix (because all traces are uploaded to the same project by default). This PR also makes sure that the trace library is fetched and built when the GDK is downloaded during CI.