[UNR-2372] Do not reset shadow data for the entire object on every update (#1788)
We previously reset the entire object's shadow data in `USpatialActorChannel::PreReceiveSpatialUpdate`. This was mostly wasted effort. With this change, we only reset the shadow data for the updated property.

This should keep parity with Unreal, as the worrisome scenario was this:
1) Let's say replicated X starts at 1 on both the server and client
2) Client changes X locally to 10
3) Server changes X locally to 10
4) Client receives replicated X from the server, but it's the same as the local value so a RepNotify wouldn't be fired on spatial
It turns out, this RepNotify also won't fire on native, because native copies the current value of the property to the shadow data before receiving the property from the wire.

* Try not resetting shadow data on receive

* Adapt for 4.23

* Update repnotify condition to match native

* Only copy if not resetting shadow data

* Make code more readable

* Remove extra newline

* Missed replace

* Replace incorrect check to initialize shadow data

* Fix indentation