[UNR-2887] Fix: Dynamic subobjects attached near actor creation would not be able to replicate their properties (#1806)
Previously, if trying to create and attach a dynamic subobject to an actor near its creation point, the fields of the dynamic component would not replicate properly after initial creation.
The reason for this was that we were not setting component write ACLs properly for these subobjects. This was due to the change made earlier, to resolve dynamic components earlier, which caused the code trying to check if the subobject is dynamically attached to think it was not dynamically attached, since it has already been resolved. This is fixed by decoupling the writing of ACLs from the resolution of the dynamic object in `SpatialSender::CreateEntity`.

Also a drive-by fix for $lld spam in logs.
Also a drive-by fix for authority queuing not checking which component we got authority over and potentially firing component updates for objects we did not have authority over.

* Initial fix by modifying write ACLs anytime we create a component

* Replace RemoveCurrent with RemoveAtSwap

* Change back to iterator, as we need to preserve order of component updates