Feature/add player disconnect command (#2715)
Add the ability for the GDK to disconnect clients whose login request is rejected in `GameMode::PreLogin`. Previously the player would be prevented from spawning but their client connection would not be terminated, resulting in them existing in a limbo state where they would sit at world origin in the target map.