[UNR-5171] Fix a functional test is already running error (#3214)
Sometimes on subsequent runs of multiworker tests we would encounter a functional test is already running error. We had a race condition between the RPC CrossServerRunTest and OnReplicated_bFinishedTest, which would prematurely and incorrectly end a test when OnReplicated_bFinishedTest=false arrived after CrossServerRunTest. (as bFinishedTest is set to false in PrepareTest)

This fixes that.