[UNR-5576] Call repnotifies in ascending RepIndex order (#3234)
In Spatial, we apply data in the order of the components (e.g. DataComponent, OwnerOnlyComponent, etc.).
This is a parity issue.

This PR changes it to call repnotifies in order of ascending repindex (which ends up being the same as native calling in ascending index of cmd in cmds array)

It also changes it so RepNotifies on actors are always called before those on subobjects. (as in native repnotifies are called first on a parent then on its subobjects.)